Move matching logic (mostly) out of ast_exec
